Simplify executor's use of notifiers
This commit removes a level of indirection for the execution task
completion notifications.  The notifier to be sent directly to the
execution adapter, instead of indirectly via a sender.

This removes the additional thread needed to receive the notifications
and forward them to the notifier.

It also removes the callback for removing the reader from the list in
the executor, though this was not being actively executed and also
contained a dead-lock.

The removal of this thread, and the simplification improves the CPU
utilization of the executor dramatically.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>