Add scheduler test for unexpected notification
Adds a test that verifies a scheduler returns an
`UnexpectedNotification` error when a transaction execution notification
is sent to it for a transaction that it was not waiting for.

Also updates the serial scheduler to properly report this error when it
doesn't have a current transaction; previously, it would stop the
background thread if it encountered this error.

Signed-off-by: Logan Seeley <seeley@bitwise.io>